refactor(portfolio): 포폴 생성 로직
